component ui adjustment
