+++ src/App.vue
@@ -1,15 +1,5 @@
 <template>
   <v-app>
-    <v-app-bar
-      app
-      color="primary"
-      dark
-    >
-
-      <v-toolbar-title>CORONA INFORMATION BOTSWANA</v-toolbar-title>
-      <v-spacer></v-spacer>
-
-    </v-app-bar>
 
     <v-content>
       <router-view></router-view>
@@ -19,14 +9,14 @@
 
 <script>
 
-import Map from './components/map';
+
 
 
 export default {
   name: 'App',
 
   components: {
-    Map,
+    
   },
 
   data: () => ({
