+++ src/components/Home.vue
@@ -1,34 +1,13 @@
 <template>
-    <header>
-        <h1 v-on:click="changeTitle">{{ title }}</h1>
-    </header>
+    <v-app>
+        
+    </v-app>
 </template>
 <script>
 export default {
-    props: {
-      title: {
-        type: String,
-        required: true
-      }
-    },
-    data(){
-        return{
-        }
-    },
-    methods: {
-      changeTitle: function(){
-        this.$emit('changeTitle', 'Vue Ninjas');
-      }
-    }
+
 }
 </script>
 <style scoped>
-header{
-    background: lightgreen;
-    padding: 10px;
-}
-h1{
-    color: #222;
-    text-align: center;
-}
+
 </style>
